After testing a few more files, inc. some from ORA GNN.
Added a few new feature test entities:
> 	-- The GO element used in ORA GNN. What is this??? -->
>
> <!ENTITY % HTML.aEndOmissable "IGNORE"
> 	-- infer </A> tags, as in ORA GNN stuff -->
>
> <!ENTITY % HTML.isindexAtEnd "IGNORE"
> 	-- allow ISINDEX after HEAD and BODY, as in ORA GNN stuff -->

NEXTID is "on" by default -- it doesn't hurt anything, I guess.

ISINDEX is allowed in %body-content if HTML.forms is "on"
@@ Hmmm... this means one must search the whole doc, not just
the head, to see if it's an index. Bad.

Expanded the BLOCKQUOTE content model to include all sorts
of %block stuff -- not just P and ADDRESS.

Fixed missing * in FORM content model.